% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{s301gradedist}
\alias{s301gradedist}
\title{S301 Grade Distributions}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 71 rows and 34 columns.}
\usage{
s301gradedist
}
\description{
A dataset containing the grade distributions for each section of STAT-S301 between 2015-2019.
}
\keyword{datasets}
